ffmpeg: re-organize Cgo module
extras.c/.h - Functions unrelated to main transcode pipeline like segmenting
and stream validation

transcoder.c/.h - Main transcode loop and other setup functions

decoder.c/.h - Input-to-raw-frames i.e demuxing, decoding

filter.c/.h - Filtering raw frames i.e. scaling, FPS changes etc

encoder.c/.h - Raw-frames-to-output i.e. encoding, muxing
